Move core ByteTracker implementation to its own RangeMap module
