+++ common/src/resources/mod.rs
@@ -1,2 +1,3 @@
 pub mod openai;
 pub mod transpose;
+pub mod range_map;
